[gitref/pullreqs/513] Render Markdown in hover info with Kakoune markup
Translate Markdown to Kakoune's markup syntax.
With inspiration from #337.

Utilize a stack to keep track of nested tags applying faces on top of
each other. When closing a nested tag, the previous face needs to be
used, rather than the default one.

In some situations the last tag will be closed with a "{default}" face.
That prevents whitespace trimming in lsp-show-hover, so remove it.

Fixes #73, #476